dev work

+++ app/database/migrations/2023_05_04_224222_create_invoices_line_items.php
@@ -18,7 +18,7 @@ class CreateInvoicesLineItems extends Migration
             $table->boolean('is_recurring')->default(FALSE); // recurring option for charges like DID rentals or membership fees
             $table->string('key_name')->nullable(); // call_costs, recording_costs, etc...
             $table->string('name')->nullable(); // Call costs, Recording costs, etc...
-            $table->float('cents')->default(0); // number of cents charged
+            $table->integer('cents')->default(0); // number of cents charged
             $table->integer('invoice_id')->nullable()->unsigned();
             $table->foreign('invoice_id')->references('id')->on('users_invoices')->onDelete('SET NULL');
